Move "image_export" to graph/export.go
